requirements.txt
altair==4.2.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clickhouse-connect==0.11.0
cryptography==46.0.5
dnspython==2.8.0
entrypoints==0.4
great_expectations==1.12.3
greenlet==3.3.2
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lz4==4.4.5
MarkupSafe==3.0.3
marshmallow==3.26.2
mistune==3.2.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.19.1
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.46
toolz==1.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
zstandard==0.25.0
